ad_manager_back/requirements.txt
asgiref==3.11.1
colorama==0.4.6
decorator==5.2.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.4.3
pillow==11.3.0
proglog==0.1.12
PyJWT==2.11.0
python-dotenv==1.2.2
sqlparse==0.5.5
tqdm==4.67.3
tzdata==2025.3
